move job order item to models
